+++ src/widget/push_button_widget.rs
@@ -31,6 +31,7 @@ pub struct PushButtonWidget {
     base_widget: BoxWidget,
     text_widget: TextWidget,
     active: bool,
+    widget_id: i32,
 }
 
 impl PushButtonWidget {
@@ -47,6 +48,7 @@ impl PushButtonWidget {
             base_widget: BoxWidget::new(),
             text_widget,
             active: false,
+            widget_id: 0,
         }
     }
 
@@ -134,6 +136,14 @@ impl Widget for PushButtonWidget {
         }
 
         None
+    }
+
+    fn set_widget_id(&mut self, widget_id: i32) {
+        self.widget_id = widget_id;
+    }
+
+    fn get_widget_id(&mut self) -> i32 {
+        self.widget_id
     }
 
     fn draw(&mut self, c: Context, g: &mut GlGraphics, clip: &DrawState) {